Added open-sans npm package
I noted the default template which I started off from uses the open-sans font, but we didn't include that font in our packaging. Now we do. This will not affect the visual appearance of the app _one bit_ yet, since there is no text being rendered, but I have a subsequent PR where that kind of stuff is coming.
