+++ .travis.yml
@@ -1,6 +1,7 @@
 sudo: false
 env:
-  - TRAVIS_NODE_VERSION="7"
+  global:
+    - TRAVIS_NODE_VERSION=7
 
 before_install:
   - rm -rf ~/.nvm && git clone https://github.com/creationix/nvm.git ~/.nvm && (cd ~/.nvm && git checkout `git describe --abbrev=0 --tags`) && source ~/.nvm/nvm.sh && nvm install $TRAVIS_NODE_VERSION
@@ -14,7 +15,7 @@ install:
 script:
   - au build
 
-cache:
-  bundler: true
-  directories:
-  - node_modules
+# Seems to cause build failures.
+#cache:
+#  directories:
+#  - node_modules
