+++ aurelia_project/tasks/build.ts
@@ -4,11 +4,12 @@ import transpileElectronApp from './transpileElectronApp';
 import processMarkup from './process-markup';
 import processCSS from './process-css';
 import { build } from 'aurelia-cli';
-import prepareFontAwesome from './prepare-font-awesome';
+import prepareFontAwesome from './prepare-fonts';
 import * as project from '../aurelia.json';
 
 export default gulp.series(
   readProjectConfiguration,
+
   gulp.parallel(
     transpile,
     transpileElectronApp,
