--- aurelia_project/tasks/prepare-font-awesome.ts
@@ -1,18 +1,0 @@
-import * as gulp from 'gulp';
-import * as merge from 'merge-stream';
-import * as changedInPlace from 'gulp-changed-in-place';
-import * as project from '../aurelia.json';
-
-export default function prepareFontAwesome() {
-  const source = 'node_modules/font-awesome';
-
-  const taskCss = gulp.src(`${source}/css/font-awesome.min.css`)
-    .pipe(changedInPlace({ firstPass: true }))
-    .pipe(gulp.dest(`${project.platform.output}/css`));
-
-  const taskFonts = gulp.src(`${source}/fonts/*`)
-    .pipe(changedInPlace({ firstPass: true }))
-    .pipe(gulp.dest(`${project.platform.output}/fonts`));
-
-  return merge(taskCss, taskFonts);
-}
