+++ aurelia_project/tasks/watch.ts
@@ -12,7 +12,6 @@ let refresh = gulp.series(
 );
 
 let watch = function () {
-
   console.log(`Watching... Ctrl+C to terminate.`);
 
   gulp.watch(project.transpiler.source, refresh).on('change', onChange);
