+++ index.html
@@ -4,6 +4,7 @@
 <head>
     <title>noTime</title>
     <link rel="stylesheet" href="output/css/font-awesome.min.css">
+    <link rel="stylesheet" href="output/css/open-sans.css">
 </head>
 
 <script>
