+++ package.json
@@ -16,6 +16,7 @@
         "electron": "^1.4.3",
         "font-awesome": "^4.7.0",
         "i18next-xhr-backend": "^1.2.0",
+        "npm-font-open-sans": "^1.0.2",
         "requirejs": "~2.3.2"
     },
     "peerDependencies": {},
