+++ src/app.scss
@@ -1,4 +1,5 @@
 @import 'styles/fonts';
+
 html {
     font-family: 'Open Sans';
     overflow: hidden;
