+++ src/styles/_fonts.scss
@@ -2,9 +2,6 @@
   font-family: 'Open Sans';
   font-style: normal;
   font-weight: 400;
-  src: url(./resources/fonts/Open-Sans.eot);
-  src: local('Open Sans'), 
-  local('OpenSans'), 
-  url(./resources/fonts/Open-Sans.woff) format('woff'), 
-  url(./resources/fonts/Open-Sans.ttf) format('truetype')
+  src: local('Open Sans'),
+  local('OpenSans'),
 }
